requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pypdf==3.17.4
reportlab==4.0.7
Pillow==10.1.0
PyJWT==2.8.0
